mavlink: move AUTOPILOT_STATE_FOR_GIMBAL_DEVICE to separate stream header
